bestinclickbackend/requirements.txt
Django>=5.0.1,<5.1.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
psycopg2-binary>=2.9.9
celery>=5.3.4,<6.0.0
redis>=5.0.1,<6.0.0
python-decouple>=3.8
Pillow>=10.1.0
django-cors-headers>=4.3.1,<5.0.0
django-filter>=23.5,<24.0.0
spacy>=3.7.0,<4.0.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.1.0,<3.0.0
numpy>=1.25.0,<2.0.0
nltk>=3.8.0,<4.0.0
openai>=1.6.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
uuid>=1.30
django-extensions>=3.2.0,<4.0.0
django-debug-toolbar>=4.2.0,<5.0.0
coverage>=7.3.0,<8.0.0
factory-boy>=3.3.0,<4.0.0
Faker>=22.0.0,<23.0.0
